* configure.in: Test for speed_t being present
        * sys_term.c (addarg): When reallocing memory, ensure in size of
        pointer. Use speed_t instead of int for termspeeds.

        * authenc.c, ext.h, state.c, utility.c: Unsigned/signed cleanup.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@15012 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/appl/telnet/telnetd/authenc.c
@@ -44,7 +44,7 @@ net_write(str, len)
 {
 	if (len < 0)
 		return 0;
-	return netwrite(str, len);
+	return netwrite(str, (size_t) len);
 }
 
 void
